add new message: class selector / wrong name
If the user tries a class selector but does not use 'apples' or 'bananas' the new message will be triggerd.
However, at the current state, once a valid selector is used, the message can't be triggered anymore.

i.e.

.apples{}
.anothername{}

The above code will not trigger the message to appear, because the grader detects '.apples'.